bucket name env variable fixed

+++ jobs/cleanup.js
@@ -9,6 +9,10 @@ schedule.scheduleJob("0 * * * *", async () => { // Runs every hour
 
   for (const cl1p of expiredCl1ps) {
     for (const file of cl1p.files) {
+      if (!bucketName || !file) {
+        continue;
+      }
+      
       await s3.deleteObject({ Bucket: bucketName, Key: file }).promise();
     }
     await Cl1p.deleteOne({ _id: cl1p._id });
